Fix the sample Nginx template for Docker in Review Board 6.
Review Board 6's documentation update for Docker broke the default Nginx
template, causing a default install to fail to load static media.

The content was copied from the sample documented Nginx template used
for a Review Board site install, but the paths were wrong for a Docker
install.

This change updates the paths to point to the right location.

Testing Done:
Compared this to the working configuration in Review Board 5 and my
test copy of the docker-compose file.

Reviewed at https://reviews.reviewboard.org/r/13379/