Remove the extra margin above the Unified Banner's changedesc field.
The change description field in the Unified Banner had a bit of extra
margin at the top, due to the default styling of text editors. This
caused the field to jump down a bit when opening, and jump back up when
closing.

We now remove that margin, keeping the field's position fixed regardless
of whether it's open or closed.

Testing Done:
Edited the field and saw that it no longer jumped.

Reviewed at https://reviews.reviewboard.org/r/13395/