Fix an issue logging connection information in upgrade failure logs.
When `rb-site upgrade` fails to upgrade a database, it outputs a lot of
debug information to a log file in order to help diagnose the issue.
Part of this process logs connection information for the database as a
JSON string. This doesn't always work, as the data there is not
guaranteed to be JSON-compatible. This failure is sandboxed and does not
affect the rest of the log, but does mean we don't always get useful
information here.

This change switches this to log a `pformat()` of the connection
information, rather than a JSON log, in order to guarantee successful
logging.

Testing Done:
Tested logging the information. Saw that the log was successful.

Reviewed at https://reviews.reviewboard.org/r/13400/