Fix bug with review request becoming a draft when a file attachment's thumbnail or extra data gets updated.
Currently, when a file attachment's extra data and/or thumbnail gets updated,
the review request that its attached to becomes a draft. This is not desired
behaviour, because these are rather "behind the scenes" updates which should
not affect the draft state of the file attachment or the review request. This
change rearranges things so that if only the extra data and/or thumbnail of
a file attachment are being updated, it does not fetch/create the review
request draft. If the caption and/or pending deletion status are updated,
this will still fetch/create the review request draft.

Testing Done:
- Triggered this bug using Power Pack's upcoming Office document review
  feature and noticed it was fixed.
- Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/13423/