Force the page body to always occupy at least 100% of the viewport.
We've had a long-standing issue where short pages sometimes displayed in
an ugly way, with the page background color cutting off just below the
content instead of filling the entire window.

This change adds the necessary rules to force the HTML body to expand to
at least the size of the viewport.

Testing Done:
Created an empty review request and saw that the light grey background
covered the entire window instead of having an ugly white block at the
bottom of the page.

Reviewed at https://reviews.reviewboard.org/r/13443/