Fix batch URL paths when using local sites.
The fix for including the SITE_ROOT for the batch URL didn't properly
include the local site (since `RB.apiCall` only uses the `prefix`
argument when `url` is undefined, in which case it assembles a URL
relative to the API root). This change fixes that up to properly include
it in the URL.

Testing Done:
Was able to publish and use the batch operations on the dashboard from
within a local site.

Reviewed at https://reviews.reviewboard.org/r/13445/