Fix up UI problems in New Review Request pre-commit workflow.
The web-based pre-commit workflow (uploading diff files) was broken in
6.0, where a diff could be selected and then after that, the rest of the
steps wouldn't work. We didn't notice that this was broken because we
had only tested with git repositories, which don't have any additional
steps (in the success case).

The problem here is that our template used XHTML syntax for closing some
childless `<div>` tags. I'm not sure exactly where this changed (either
underscore or jquery), but that no longer works. This meant that our
elements for many of the pages were improperly nested.

Testing Done:
Went through the pre-commit new review request flow with an SVN
repository and saw that I was able to upload a diff to create a review
request.

Bugs Fixed: 5013

Reviewed at https://reviews.reviewboard.org/r/13457/