Prevent empty reviews from being published.
With the change to "Create Review" in the new unified banner, we made it
much easier to have an empty review get published. It used to be that
closing the review dialog with an empty review would delete the draft,
but we now keep it around.

This change adds a validation step for review publishing to ensure that
the review is not empty.

Testing Done:
- Attempted to publish an empty review from both the unified banner and
  the review dialog. Saw the expected error message.
- Ran Python and JS unit tests.

Reviewed at https://reviews.reviewboard.org/r/13461/