Make inline rendering explicit for review UI diff fragments.
The render call for review UIs when rendering as a fragment for the diff
viewer was relying on the default value of the `inline` parameter, which
confused me for a while. This change makes that value explicit in the
function call.

Testing Done:
Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/13522/