Download original versions of binary files for diffs.
When viewing a diff where we have a file attachment for the new version
of a binary file, this change will download the original file and create
a matching file attachment.

Testing Done:
Created a file attachment for the new version of an image in a diff.
Opened the diff viewer and saw that the original file was downloaded,
stored in a file attachment, and the image review UI showed a diff in
the diff viewer.

Reviewed at https://reviews.reviewboard.org/r/13582/