Detect and store the MIME type of files in a diff.
This change adds a step to detect the MIME type for the source and dest
files in a FileDiff. This gets stored in the `FileDiff.extra_data`
field.

Testing Done:
Loaded a diff and saw that the MIME types were correctly detected and
stored.

Reviewed at https://reviews.reviewboard.org/r/13582/