Remove caching for HideActionHook results.
The action implementation was caching the result of evaluating all of
the registered HideActionHook instances, but that ends up potentially
giving bad results. If an extension is being enabled or disabled, the
cached value can be incorrect. This is particularly noticable in the
devserver, but could affect production deployments too.

Testing Done:
Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/13509/