Convert reviewable views to TypeScript/spina.
This change converts the reviewable views (and associated utility views)
to TypeScript/spina.

While I was here I disconvered that one of the tests for DiffReviewable
was intended to be async, but didn't take a `done` callback and was
therefore skipping the meat of the test. After fixing that, it became
evident that the implementation of the test was broken, so it's been
rewritten.

Testing Done:
- Ran js-tests.
- Smoke tested reviewing diffs and file attachments.

Reviewed at https://reviews.reviewboard.org/r/13531/