Remove half-baked inline-actions-header from Review UI diff HTML.
When review UIs are rendered into the diff viewer, the fragment template
was including an "inline-actions-header" table row. In a case that in
practice would never be hit, this could theoretically contain "Review"
and "New Comment" buttons (like we have in the file attachment
thumbnail), but those buttons had no handlers. In the more common case
of a review UI that renders inline review content, the result was an
ugly table row that didn't have anything in it, but took up space.

Testing Done:
Looked at the diff viewer with image review UIs rendered and saw that
there weren't any ugly gaps.

Reviewed at https://reviews.reviewboard.org/r/13535/