Defer callback after images are loaded in image review UIs.
The callback that handles the initial sizing/scaling for images was
sometimes running before jquery finished setting the `initial-width` and
`initial-height` data, which could result in one of the two images in
the two-up view being scaled and one not. This change adds a `_.defer`
call to unwind the stack before we trigger that callback.

Testing Done:
Loaded a diff that had a bunch of changed images with @2x versions. Saw
that I was no longer able to trigger the case where one out of the two
images was full size and one was 50%.

Reviewed at https://reviews.reviewboard.org/r/13545/