Make some layout tweaks to the two-up mode for image diffs.
The two-up mode was displaying the images as floating with the
vertical-align being the baseline, but this didn't match the way the
other diff modes aligned things (at the top). This changes the layout to
use flexbox to achieve the same layout but have things aligned at the
top of the container instead of the bottom.

Testing Done:
Looked at an image diff.

Reviewed at https://reviews.reviewboard.org/r/13575/