Make a bunch of improvements to the FileAttachment admin list.
The way attachments were listed in the admin site was very difficult to
use. The `file` field was linking to the download URL, and the only
other link to the change page was the caption, which could be empty.

This change improves the admin list so that the filename rather than the
path is listed (potentially cropped to preserve the page with), but that
always links to the change page. The "Review Request ID" column now
properly shows the linked review request even when the attachment is
part of a draft or in the inactive list. A new column has been added
with the download link.

Testing Done:
Used the admin list and was actually able to get to the change page for
many of my attachments.

Reviewed at https://reviews.reviewboard.org/r/13576/