Fix duplicate files showing up in review request emails.
For a long time, I've noticed that filenames are duplicated in the
review request emails. Having had my head in the commit code, I finally
realized why: we're iterating over `DiffSet.files.all()`, which includes
both the files in the commit(s) as well as the files in the cumulative
diff.

This change fixes it so we iterate over the `cumulative_files` property,
which shows only the FileDiffs in the cumulative diff.

Testing Done:
Looked at the review request email previews and no longer saw duplicated
filenames.

Reviewed at https://reviews.reviewboard.org/r/13578/