Add documentation on the Review Board Installer.
This updates our documentation to rework the guidance for installing
Review Board to prioritize the new Review Board Installer.

The new page for this covers the system requirements, running the
installer, advanced options for unattended installs and proxies, and
then guiding to the appropriate place for completing a site
installation.

A page listing all compatible Linux versions is available, and linked to
on the installer and old Linux pages.

The Linux, Windows, and macOS pages have been updated to reference the
new installer docs as well.

The database versions in the new and existing compatibility lists have
also been bumped up to modern versions that we'll want to push going
forward, based on Django's deprecation cycle.

This should hopefully do a lot to help new users get started with
Review Board.

Testing Done:
Built the docs. Checked for build errors, spelling errors, and bad
links.

Read through it several times, looking for issues or confusion in
the new docs. It needs another pair of eyes.

Reviewed at https://reviews.reviewboard.org/r/13508/